Lift the actions handler, state gate, and actions section to shared homes
