+++ src/components/device/automation-editor/api-action-editor.ts
@@ -35,7 +35,6 @@ import { renderMarkdown } from "../../../util/markdown.js";
 import { registerMdiIcons } from "../../../util/register-icons.js";
 import { scrollFlashRow } from "../field-highlight.js";
 import { fieldHighlightStyles } from "../field-highlight.styles.js";
-import "./automation-action-list.js";
 import {
   actionsFocus,
   entryFieldFocus,
@@ -43,11 +42,11 @@ import {
   paramFocus,
 } from "./automation-focus.js";
 import { CallableAutomationEditor } from "./callable-editor.js";
+import { renderActionsSection } from "./render-actions-section.js";
 import "./callable-params-editor.js";
 import { emptyAutomationTree } from "./serialise.js";
 
 import "@home-assistant/webawesome/dist/components/icon/icon.js";
-import "@home-assistant/webawesome/dist/components/spinner/spinner.js";
 
 registerMdiIcons({
   "open-in-new": mdiOpenInNew,
@@ -88,15 +87,8 @@ export class ESPHomeApiActionEditor extends CallableAutomationEditor<ApiActionLo
   }
 
   protected render() {
-    if (this._loading) {
-      return html`<div class="ae-empty">
-        <wa-spinner></wa-spinner>
-        ${this._localize("device.loading_automation_catalog")}
-      </div>`;
-    }
-    if (this._parseError.active) {
-      return this._parseError.renderPanel(this._localize);
-    }
+    const gate = this.renderStateGate();
+    if (gate) return gate;
     const automation = this.value ?? emptyAutomationTree();
     const devices = this._available?.devices ?? [];
     const scripts = this._available?.scripts ?? [];
@@ -116,25 +108,20 @@ export class ESPHomeApiActionEditor extends CallableAutomationEditor<ApiActionLo
         .namePlaceholder=${this._localize("device.api_action_variable_name_placeholder")}
         @value-change=${this._onVariablesChange}
       ></esphome-callable-params-editor>
-      <div class="field">
-        <label class="field-label"> ${this._localize("device.automation_action")} </label>
-        <p class="field-description">
-          ${renderMarkdown(this._localize("device.api_action_actions_description"))}
-        </p>
-        <esphome-automation-action-list
-          no-header
-          .focusTarget=${actionsFocus(focus)}
-          .actions=${automation.actions}
-          .catalog=${actions}
-          .conditionCatalog=${conditions}
-          .scripts=${scripts}
-          .devices=${devices}
-          .board=${this.board}
-          .yaml=${this.yaml}
-          ?disabled=${disabled}
-          @actions-change=${this._onActionsChange}
-        ></esphome-automation-action-list>
-      </div>
+      ${renderActionsSection({
+        automation,
+        catalog: actions,
+        conditionCatalog: conditions,
+        scripts,
+        devices,
+        board: this.board,
+        yaml: this.yaml,
+        disabled,
+        localize: this._localize,
+        focusTarget: actionsFocus(focus),
+        descriptionKey: "device.api_action_actions_description",
+        onActionsChange: this._onActionsChange,
+      })}
       ${this.renderFooter({
         label: this._localize("device.delete_api_action"),
         message: (location) =>
@@ -230,11 +217,6 @@ export class ESPHomeApiActionEditor extends CallableAutomationEditor<ApiActionLo
         variables: e.detail.value,
       },
     });
-  };
-
-  private _onActionsChange = (e: CustomEvent<{ actions: AutomationTree["actions"] }>) => {
-    e.stopPropagation();
-    this._engine.withValue({ actions: e.detail.actions });
   };
 }
